fix(backend): validaciones, integridad de inventario y motor de retenciones
Revisión de backend (bloques 1 y 2). Suite 123/123 en verde.

- Validación de montos/cantidades/% en schemas (gt=0, ge=0, 0-100)
- Guard de sobreventa en confirmar_venta (400 antes de descontar stock)
- Elimina N+1 en list_ventas/get_venta con selectinload + helper
- Numeración robusta unificada (core/numbering.py): MAX + parseo tolerante
- Dashboard: excluye ventas ANULADAS en "ventas por marca"
- Stock fraccionario: Numeric(12,3) en producto y kardex, sin redondeo
- Admin del seed configurable por env var (SEED_ADMIN_*)
- Auto-CxC al confirmar venta (espejo de compras->CxP)
- Retenciones híbridas: perfil tributario del cliente + tarifas de
  ParametroTributario + tope en UVT + override manual por factura; fija reteica
- Script de migración SQLite para columnas de retención en clientes
- Tests nuevos (test_validaciones) y actualización de casos afectados

+++ backend/app/modules/inventario/models.py
@@ -36,8 +36,8 @@ class MovimientoInventario(Base):
     tipo = Column(SAEnum(TipoMovimientoInventario), nullable=False)
     origen = Column(SAEnum(OrigenMovimiento), nullable=False)
     cantidad = Column(Numeric(12, 3), nullable=False)
-    stock_antes = Column(Integer, nullable=False)
-    stock_despues = Column(Integer, nullable=False)
+    stock_antes = Column(Numeric(12, 3), nullable=False)
+    stock_despues = Column(Numeric(12, 3), nullable=False)
     costo_unitario = Column(Numeric(18, 2), nullable=True)
 
     # FK lógicas (sin constraint real) — mismo patrón que CuentaPorPagar.compra_id
